realized an issue with the int to float conversion

+++ autoSimple.py
@@ -9,7 +9,7 @@ x = locate(.05)
 while copy:
     for i in x:
         auto.click(i)
-        time.sleep(int(bigT))
+        time.sleep(float(bigT))
     if key2.is_pressed('c'):
         copy = False
         break
